SDK. Check transactions

+++ wavesplatform/src/main/java/com/wavesplatform/sdk/model/request/node/BurnTransaction.kt
@@ -29,7 +29,7 @@ class BurnTransaction(
             Bytes.concat(
                     byteArrayOf(type.toByte()),
                     byteArrayOf(version.toByte()),
-                    byteArrayOf(chainId.toByte()),
+                    byteArrayOf(chainId),
                     Base58.decode(senderPublicKey),
                     Base58.decode(assetId),
                     Longs.toByteArray(quantity),
